some gardens have LARGE collections, taking minutes to process
django handles stuff lazy, and we need to respect that, that's all.  so when
we get a query-set, we don't process it all at once, we rather send it back
in chunks.  interaction is a bit complex, based on tokens and I'm afraid
there might be synchronization problems, but the worst that can happen is
that you get old results in your new query.