adding unused class for vernacular names
this is to be linked to taxon, a language, and a geographic area.  in
perspective it should let us specify that 'cidra' is the Spanish name for
Sechium edule in the area around Medellín, and that the same Sechium edule,
still in Spanish, is called different names, among which 'chayote', in
different areas, and that the same name 'chayote' is also used in English,
globally.